I added functionality to the questions controller to fix the bug
I added logic to the controller questions_controller.rb in order to fix
the bug if when a user deletes a question, if a content set contains
only this question, the corresponding PostGres table (question_contents)
is not updated (leaving the reference to the question when the question
is in fact deleted). This will leave 0 quesitons in the content set,
which will cause the Brain Race game to crash if this content set it
selected. Now, when a user deletes a question, all references to this
question will be removed from the table question_contents and will mark
all content sets with no questions as inactive so it does not show up in
the Brain Race game. NOTE: This solution is a poor solution (N+1)
problem. But since I do not know how to do this query in the confines of
this RoR app and the ActiveRecord, I had to get something working ASAP.
The proper way to do this would be to modify the code in the following
models content_set.rb, question.rb and question_content.rb and also
modyfy the code in the following controllers questions_controller.rb and
content_sets_controller.rb  (and possibly add a new controller for the
question_contents) in order to handle this query operation more
eloquently. The following link contains an example similar to what I
need to do in this app, but I just need time to learn RoR and implement
the more eloquent solution.

http://guides.rubyonrails.org/v2.3.11/association_basics.html